Files open asynchronously, removed disable new file on open always.
The file that is read that contains the projects files is read asyncly
instead of syncly. This should help startup time a bit.

The `Disable New File On Open Always` setting had to be removed because
of this. The new file was being opened before the files were loaded in,
so it doesn't work correctly anymore. That setting is now just how
`Disable New File On Open` works.