First steps towards saving scroll position
Atom's api for this seems very lacking, but I've done what I can for
now. Whenever files are saved, the scroll position is saved, not when
the user scrolls. This is done for fear of performance issues until the
package is optimized better. Atom also doesn't scroll if the specified
line is in the view and it doesn't do anything special to get the line
in the same spot it was before. This creates a kind of weird
restoration, but it's better than nothing (I think).