New windows will no longer restore a project when they are opened
This will fail if a window is opened, closed and immediately after,
another window is opened. If any other session editing interaction
happens, it should work fine.

I'm not sure how to write tests for some of these things, but I edited
the tests so that they pass and test what I know how to test.