Restore over changes if cached contents unchanged
When the disk image of a buffer differs from the cachedDiskContents in
atom, we skip replacing the contents with our own version because,
presumably, the changed disk image is the one the user wants.  But it's
possible some other buffer-recovery function has resulted in a cache
difference (perhaps in core atom) and we are fooled because we look at
the cache when saving but at the disk when restoring.

Look at the cache when restoring so it matches what we saw when saving
since the check is only interested in ensuring we do not discard changes
from some outside process.

Fixes #129